The runtime deps can be found via pkginfo --runtimedepfiles <path> This allow us to involve pkgmk in the business. It add automatically the runtime deps before packaging
Les dépendances runtime peuvent être trouvées via pkginfo --runtimedepfiles <chemin>
Cela nous permet de demander à pkgmk d'ajouter automatiquement les dependances runtime avant l'empaquetage

+++ elf.h
@@ -42,9 +42,9 @@ typedef struct elf_internal_ehdr_32b {
 	unsigned int     e_shoff;
 } Elf_Ehdr_32Bit;
 typedef struct elf_internal_ehdr_64b {
-   unsigned long long    e_entry;
-   unsigned long long    e_phoff;
-   unsigned long long    e_shoff;
+   long    e_entry;
+   long    e_phoff;
+   long    e_shoff;
 } Elf_Ehdr_64Bit;
 
 typedef struct elf_internal_ehdr_end {
